HW-4 Add MovieList abstract layer to share common functions between Movie and Recent - Move movieAdapter from Base to MovieList(https://github.com/StudyFork/GoogryAndroidArchitectureStudy/pull/915#discussion_r530954343)
