+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/ui/contract/BaseContract.kt
@@ -1,6 +1,5 @@
 package com.example.googryandroidarchitecturestudy.ui.contract
 
-import com.example.googryandroidarchitecturestudy.domain.Movie
 import com.example.googryandroidarchitecturestudy.domain.UrlResource
 
 interface BaseContract {
@@ -13,11 +12,9 @@ interface BaseContract {
         fun showQueryEmpty()
         fun showNoSearchResult()
         fun showSearchFailed(e: Exception)
-        fun showMovieList(items: List<Movie>)
     }
 
     interface Presenter {
         fun selectUrlItem(item: UrlResource)
-        suspend fun queryMovieList(query: String)
     }
 }
