+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/ui/contract/RecentContract.kt
@@ -3,11 +3,11 @@ package com.example.googryandroidarchitecturestudy.ui.contract
 import com.example.googryandroidarchitecturestudy.domain.RecentSearch
 
 interface RecentContract {
-    interface View : BaseContract.View {
+    interface View : MovieListContract.View {
         fun showRecentChips(recentList: List<RecentSearch>)
     }
 
-    interface Presenter : BaseContract.Presenter {
+    interface Presenter : MovieListContract.Presenter {
         suspend fun getRecentSearch()
     }
 }
