+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/ui/presenter/MoviePresenter.kt
@@ -2,10 +2,11 @@ package com.example.googryandroidarchitecturestudy.ui.presenter
 
 import com.example.googryandroidarchitecturestudy.data.repository.MovieRepository
 import com.example.googryandroidarchitecturestudy.ui.contract.MovieContract
+import com.example.googryandroidarchitecturestudy.ui.contract.MovieListContract
 
-class MoviePresenter(
-    view: MovieContract.View,
+open class MoviePresenter(
+    view: MovieListContract.View,
     repository: MovieRepository
-) : BasePresenter(view, repository), MovieContract.Presenter {
+) : MovieListPresenter(view, repository), MovieContract.Presenter {
 
 }
